+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Test/Core/FloaterTest.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using NUnit.Framework;
+using Unity.Mathematics;
 
 // ReSharper disable CheckNamespace
 namespace ChoyUtilities.Test {
@@ -50,7 +51,31 @@ namespace ChoyUtilities.Test {
             var floater = new Floater(insert);
             Assert.AreEqual(insert, floater);
         }
+
+        [Test]
+        public void Floater_Float3ArrayTest() {
+            var insert = new float3[] {
+                new float3(1, 2, 3),
+                new float3(4, 5, 6),
+                new float3(7, 8, 9)
+            };
+            var expected = new[] {
+                1, 2, 3,
+                4, 5, 6,
+                7, 8, 9
+            };
+            var floater = new Floater(insert);
+            Assert.AreEqual(expected, floater);
+        }
         
-        
+        [Test]
+        public void Floater_CharArrayTest() {
+            var insert = new [] {
+                'H', 'e', 'l', 'l', 'o', ' ', 'W', 'o', 'r', 'l', 'd', '!',
+            };
+            const string expected = "Hello World!";
+            string floater = new Floater(insert);
+            Assert.AreEqual(expected, floater);
+        }
     }
 }